seperate database version update test
