+++ circle.yml
@@ -9,9 +9,10 @@ test:
         - py.test --verbose --capture=no counterpartylib/test/unit_test.py
         - py.test --verbose --capture=no counterpartylib/test/parse_block_test.py
         - py.test --verbose --capture=no --skiptestbook=all counterpartylib/test/integration_test.py
+        - py.test --verbose --capture=no counterpartylib/test/contracts_test.py
         - py.test --verbose --capture=no --skiptestbook=mainnet -k test_book counterpartylib/test/reparse_test.py
         - py.test --verbose --capture=no --skiptestbook=testnet -k test_book counterpartylib/test/reparse_test.py
-        - py.test counterpartylib/test/contracts_test.py
+        - py.test --verbose --capture=no counterpartylib/test/database_version_test.py
 machine:
     pre:
         - mkdir -p ~/.local/share/counterparty;
